chore: Update data loading to use int32 dtype in ThePileDataset

+++ src/kazeML/jax/bluejay_llm/bluejay_llm/dataloader.py
@@ -6,7 +6,7 @@ import numpy as np
 class ThePileDataset(Dataset):
     def __init__(self, path: str, max_length: int = 1024, process_id: int = 0, num_processes: int = 1):
         self.max_length = max_length
-        data = np.memmap(path, dtype=np.uint16, mode="r")
+        data = np.memmap(path, dtype=np.uint16, mode="r").astype(np.int32)
         total_length = len(data)
         data = data[process_id:total_length:num_processes]
         self.data = torch.tensor(data, dtype=torch.int32)
